Создал отдельный виджет AddTaskDialog
