+++ lib/features/home/widgets/widgets.dart
@@ -2,5 +2,6 @@
 // generated by: https://marketplace.visualstudio.com/items?itemName=elmsec.dart-flutter-exports:
 
 
+export 'add_task_dialog.dart';
 export 'tasks_list.dart';
 export 'task_item.dart';
